telegram-webapp/pyproject.toml
[project]
name = "telegram-webapp"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["python-telegram-bot>=22.1"]
